fix(blocks): update
